gentodo: gentodo/ -> src/, split files
src/:

Moving `gentodo/` to `src/` is likely the more wise decision moving
forward as it will allow for potential developers and users to find
where the source files are stored easily.

Gentodo class and functions:

Split the Gentodo class and its functions into a new file, aptly named
`gentodo.py` to reduce the size of `main.py`

The original Gentodo class is commented out temporarily so it can
undergo a code review from a colleague.